+++ generators/app/templates/ts/scripts/test.ts
@@ -1,4 +1,4 @@
-const jest = require('jest');
+import * as jest from 'jest';
 
 process.env.BABEL_ENV = 'test';
 process.env.NODE_ENV = 'test';