+++ generators/app/templates/ts/server/index.ts
@@ -1,13 +1,13 @@
-const path = require('path');
-const express = require('express');
-const history = require('connect-history-api-fallback');
-const chalk = require('chalk');
-const { devMiddleware, hotMiddleware } = require('./middleware/dev-middleware');
+import path from 'path';
+import express from 'express';
+import hist from 'connect-history-api-fallback';
+import chalk from 'chalk';
+import { devMiddleware, hotMiddleware } from './middleware/dev-middleware';
 
 const app = express();
 const port = process.env.port || 8080;
 
-app.use(history());
+app.use(hist());
 
 app.use(devMiddleware);
 app.use(hotMiddleware);