+++ CHANGELOG.md
@@ -6,8 +6,8 @@
 * Update template dependencies and tweak config. _(dzervoudakes)_
 * Update lint rules. _(dzervoudakes)_
 * Add technical requirements to README template. _(dzervoudakes)_
-* Add `babel` support for optional chaining. _(dzervoudakes)_
 * Remove forced `save-dev=true` from `.npmrc`. _(dzeroudakes)_
+* Update TypeScript template scripts, config and server syntax. _(dzervoudakes)_
 
 ## v0.1.2